+++ changelog.md
@@ -2,5 +2,6 @@
 
 ## devel
 
+* add error logging for failed yaml parses
 * add verbose flag to enable debug output
 * add current logs for all mapi container pods